+++ src/test/java/com/tomholmes/product/htmx/demo/service/ContactEmailServiceImplTest.java
@@ -25,7 +25,7 @@ public class ContactEmailServiceImplTest
     @Autowired
     private ContactEmailService service;
 
-    private final static int id = 0;
+    private final static long id = 0;
     private final static String prefix = "Mr.";
     private final static String firstname = "Thomas";
     private final static String middlename = "Joseph";
